Code overhaul
Add check to downloader to see if we already have the installation file
and that
the hash is correct.

Made sure Python is installed before we download the installer to
uninstall it again. (Do we don't waste time or bandwidth.)

Added changeInstallDir()
Added a check to see if we already have the installer to
downloadPython()
Added a 2 second pause to let the user see what directory Python is set
to be installed in.
Made the menu loop until the program is quit.